Refactored the load function using the newly created FeedImagesMapper static map function
